+++ errors/interfaces.go
@@ -1,5 +1,7 @@
 package errors
 
+import "github.com/pkg/errors"
+
 // ErrorCollection holds a list of errors.
 type ErrorCollection interface {
 	Errors() []error
@@ -13,6 +15,11 @@ type ContextualError interface {
 // Causer is the interface defined in github.com/pkg/errors for specifying a parent error.
 type Causer interface {
 	Cause() error
+}
+
+// StackTracer is the interface defined in github.com/pkg/errors for exposing stack trace from an error.
+type StackTracer interface {
+	StackTrace() errors.StackTrace
 }
 
 // Handler is responsible for handling an error.